XP-4807 Option Set - Form inputs inside itemsets are not cleared when an option is unselected - Adjusted FormOptionSetOptionView items reset process on deselect to make it iterate through the whole property set and remove all non-set properties. This forces correct clean of option set descendants and enables default values to be set. - Adjusted update() process to enforce option view state that may be affected by parent option set option select/deselect - Made little cleaning in form set classes
