+++ modules/admin/admin-ui/src/main/resources/web/admin/common/js/form/set/itemset/FormItemSetView.ts
@@ -1,11 +1,6 @@
 module api.form {
 
     import PropertySet = api.data.PropertySet;
-    import Property = api.data.Property;
-    import PropertyArray = api.data.PropertyArray;
-    import Value = api.data.Value;
-    import ValueType = api.data.ValueType;
-    import ValueTypes = api.data.ValueTypes;
 
     export interface FormItemSetViewConfig {
 
@@ -47,16 +42,6 @@ module api.form {
                 parent: this.getParent(),
                 propertyArray: this.getPropertyArray(this.parentDataSet)
             });
-        }
-
-        protected getPropertyArray(propertySet: PropertySet): PropertyArray {
-            var propertyArray = propertySet.getPropertyArray(this.formSet.getName());
-            if (!propertyArray) {
-                propertyArray = PropertyArray.create().setType(ValueTypes.DATA).setName(this.formSet.getName()).setParent(
-                    this.parentDataSet).build();
-                propertySet.addPropertyArray(propertyArray);
-            }
-            return propertyArray;
         }
     }
 }
