+++ modules/admin/admin-ui/src/main/resources/web/admin/common/js/form/set/optionset/FormOptionSetOptionView.ts
@@ -41,6 +41,12 @@ module api.form {
             this.setCheckBoxDisabled()
         }).bind(this);
 
+        private radioDeselectHandler: (event: api.data.PropertyValueChangedEvent) => void = ((event: api.data.PropertyValueChangedEvent) => {
+            if (event.getPreviousValue().getString() == this.getName()) {
+                this.deselectHandle();
+            }
+        }).bind(this);
+
         constructor(config: FormOptionSetOptionViewConfig) {
             super(<FormItemViewConfig> {
                 className: "form-option-set-option-view",
@@ -87,17 +93,7 @@ module api.form {
 
             layoutPromise.then((formItemViews: FormItemView[]) => {
 
-                if (this.isOptionSetExpandedByDefault() || this.getThisPropertyFromSelectedOptionsArray() != null) {
-                    this.expand();
-                }
-
-                if (this.isOptionSetExpandedByDefault() && this.getThisPropertyFromSelectedOptionsArray() == null) {
-                    this.disableFormItems();
-                }
-
-                if (this.getThisPropertyFromSelectedOptionsArray() != null) {
-                    this.addClass("selected");
-                }
+                this.updateViewState();
 
                 if (this.formOptionSetOption.getFormItems().length > 0) {
                     this.addClass("expandable");
@@ -219,12 +215,8 @@ module api.form {
         }
 
         private subscribeOnRadioDeselect(property: Property) {
-            var radioDeselectHandler = (event: api.data.PropertyValueChangedEvent) => {
-                if (event.getPreviousValue().getString() == this.getName()) {
-                    this.deselectHandle();
-                }
-            }
-            property.onPropertyValueChanged(radioDeselectHandler);
+            property.unPropertyValueChanged(this.radioDeselectHandler);
+            property.onPropertyValueChanged(this.radioDeselectHandler);
         }
 
         private makeSelectionCheckbox(): api.ui.Checkbox {
@@ -330,9 +322,20 @@ module api.form {
 
             var array = this.getOptionItemsPropertyArray(this.parentDataSet);
             array.getSet(0).forEach((property) => {
-                array.getSet(0).removeProperty(property.getName(), property.getIndex());
+                this.removeNonDataProperties(property);
             });
+
             this.update(this.parentDataSet);
+        }
+
+        private removeNonDataProperties(property: Property) {
+            if (property.getType().equals(ValueTypes.DATA)) {
+                property.getPropertySet().forEach((prop) => {
+                    this.removeNonDataProperties(prop);
+                });
+            } else {
+                property.getParent().removeProperty(property.getName(), property.getIndex());
+            }
         }
 
         private isSelectionLimitReached(): boolean {
@@ -356,11 +359,26 @@ module api.form {
         update(propertySet: api.data.PropertySet, unchangedOnly?: boolean): Q.Promise<void> {
             this.parentDataSet = propertySet;
             var propertyArray = this.getOptionItemsPropertyArray(propertySet);
+
             return this.formItemLayer.update(propertyArray.getSet(0), unchangedOnly).then(() => {
                 if (!this.isRadioSelection()) {
                     this.subscribeCheckboxOnPropertyEvents();
+                } else if (this.getThisPropertyFromSelectedOptionsArray() == null) {
+                    wemjq(this.getHTMLElement()).find("input:radio").first().prop('checked', false);
                 }
+
+                this.updateViewState();
             });
+        }
+
+        private updateViewState() {
+            this.expand(this.isOptionSetExpandedByDefault() || this.getThisPropertyFromSelectedOptionsArray() != null);
+
+            if (this.isOptionSetExpandedByDefault() && this.getThisPropertyFromSelectedOptionsArray() == null) {
+                this.disableFormItems();
+            }
+
+            this.toggleClass("selected", this.getThisPropertyFromSelectedOptionsArray() != null);
         }
 
         broadcastFormSizeChanged() {
